requirements.txt

streamlit>=1.32.0

groq>=0.9.0

# --- LangChain ---
langchain>=0.2.5
langchain-core>=0.2.5
langchain-groq>=0.1.5

# --- Databases ---
pymongo>=4.6.0
redis>=5.0.1

python-dotenv>=1.0.1
typing-extensions>=4.9.0

bcrypt>=4.1.2
